backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0
pydantic-core>=2.18.0,<3.0.0



# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.10
pgvector==0.4.1

# Auth
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# HubSpot 
hubspot-api-client==7.0.0

# OpenAI integration
openai==1.3.7

# Langchain 
langchain==0.3.27

# Vector database 
chromadb==0.4.18
tiktoken==0.5.1

# Document processing 
PyPDF2==3.0.1
python-docx==1.1.0
pypdf==6.0.0  # Alternative to PyPDF2, not needed

# HTTP client (required for RAG)
httpx==0.25.2
requests==2.31.0

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
